Created redux reducers and actions, refactored Nav using Redux Thunk and Redux Hooks
