+++ client/docs/whaleTable.js
@@ -1,0 +1,235 @@
+const whaleTable =[
+    {
+        "name": "Jay-Z",
+        "wallet": "0x3b417faee9d2ff636701100891dc2755b5321cc3",
+        "image": "https://dappradar.com/blog/static/e5c9498d43c1a9984ce012b24c0b01a3/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb1-jayz.png",
+        "worth": "$124,189.88",
+        "worth_num": 124189.88
+    },
+    {
+        "name": "Mike Tyson",
+        "wallet": "0x7217bc604476859303a27f111b187526231a300c",
+        "image": "https://dappradar.com/blog/static/73071f7bc5dd23b6a1d208b7f322806f/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb2-mike-tyson.png",
+        "worth": "$16,102.69",
+        "worth_num": 16102.69
+    },
+    {
+        "name": "3lau",
+        "wallet": "0xd2aff66959ee0e6f92ee02d741071ddb5084bebb",
+        "image": "https://dappradar.com/blog/static/b4d70f07cc070d5525f53c3128d30c69/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb3-3lau.png",
+        "worth": "$52,837.07",
+        "worth_num": 52837.07
+    },
+    {
+        "name": "Stephen Curry",
+        "wallet": "0x3becf83939f34311b6bee143197872d877501b11",
+        "image": "https://dappradar.com/blog/static/7dff06f69a0931b84a19f9967bf94fa9/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb4-curry.png",
+        "worth": "$322,032.97",
+        "worth_num": 322032.97
+    },
+    {
+        "name": "Steve Aoki",
+        "wallet": "0xe4bbcbff51e61d0d95fcc5016609ac8354b177c4",
+        "image": "https://dappradar.com/blog/static/fa6b60ae6f956b41c127a24583594846/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb5-aoki.png",
+        "worth": "$7.02 Million",
+        "worth_num": 7020000
+    },
+    {
+        "name": "LaMello Ball",
+        "wallet": "0xc1064e3662b0718357e9050694a3bfeaabede8ab",
+        "image": "https://dappradar.com/blog/static/163311af384bbd8ebdb5ac1a4edd5aba/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb6-lamello.png",
+        "worth": "$508,557.43",
+        "worth_num": 508557.43
+    },
+    {
+        "name": "Lindsay Lohan",
+        "wallet": "0x3781d92e5449b5b689fee308ded44882085b6312",
+        "image": "https://dappradar.com/blog/static/450e2b60e3d1304c1e46c0c04a78a37a/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb7-lindsay.png",
+        "worth": "$22,415.23",
+        "worth_num": 22415.23
+    },
+    {
+        "name": "Mark Cuban",
+        "wallet": "0xa679c6154b8d4619af9f83f0bf9a13a680e01ec",
+        "image": "https://dappradar.com/blog/static/782886aabbaa9e43359c68e5c976667e/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb8-cuban.png",
+        "worth": "$788,995.29",
+        "worth_num": 788995.29
+    },
+    {
+        "name": "Beeple",
+        "wallet": "0xc6b0562605d35ee710138402b878ffe6f2e23807",
+        "image": "https://dappradar.com/blog/static/0414dd77baead2a346cc5609fdc5d10b/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb9-beeple.png",
+        "worth": "$688,307.81",
+        "worth_num": 688307.81
+    },
+    {
+        "name": "Snoop Dog",
+        "wallet": "0xce90a7949bb78892f159f428d0dc23a8e3584d75",
+        "image": "https://dappradar.com/blog/static/3270fd75db41e9534e3b42bf341426df/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-snoop.png",
+        "worth": "$3.62 Million",
+        "worth_num": 3620000
+    },
+    {
+        "name": "Ja Rule",
+        "wallet": "0x8d3bc45d7b30013c37c141f6ce7c981b2613efaa",
+        "image": "https://dappradar.com/blog/static/8fe8bd36f5e274de8e7e8a1a3021c84a/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-serena-williams.png",
+        "worth": "$111,777.20",
+        "worth_num": 111777.2
+    },
+    {
+        "name": "Alexis Ohanian",
+        "wallet": "0x0ed1e02164a2a9fad7a9f9b5b9e71694c3fad7f2",
+        "image": "https://dappradar.com/blog/static/2f33d435fcd8286dcf9bb3faaa472592/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-alexis-ohanian.png",
+        "worth": "$2.61 Million",
+        "worth_num": 2610000
+    },
+    {
+        "name": "Mike Shinoda",
+        "wallet": "0xb55eb9bd32d6ab75d7555192e7a3a7ca0bcd5738",
+        "image": "https://dappradar.com/blog/static/066270c54fe8983b9a27d4bcb615e21a/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-celeb-mike-shinoda.png",
+        "worth": "$184,353.75",
+        "worth_num": 184353.75
+    },
+    {
+        "name": "Shaquille O Neal",
+        "wallet": "0x3c6aeff92b4b35c2e1b196b57d0f8ffb56884a17",
+        "image": "https://dappradar.com/blog/static/57742c2f905feb6b820e9b2c4e9d6433/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-shaquille-oneal-creature-world.png",
+        "worth": "$57,893.24",
+        "worth_num": 57893.24
+    },
+    {
+        "name": "Elijah Wood",
+        "wallet": "0xf6de94be96f80602d90bf29bd9e88a0e843b2eb9",
+        "image": "https://dappradar.com/blog/static/5bcda95cbb34df3111d933b4ba911b68/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-elijah-wood.png",
+        "worth": "$138,030.13",
+        "worth_num": 13830.13
+    },
+    {
+        "name": "Seth Curry",
+        "wallet": "0x9114b66e4bd387eb832d1477e86ede0cf7f76115",
+        "image": "https://dappradar.com/blog/static/ba31b70abac1841a82c63e2c7fdc0b17/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-seth-curry-celeb-nft.png",
+        "worth": "$8,972.58",
+        "worth_num": 8972.58
+    },
+    {
+        "name": "Joel Madden",
+        "wallet": "0x35cded880959f93c415723902f91f964367a4dcd",
+        "image": "https://dappradar.com/blog/static/5d2af02cd32510986127b8d5f7a8d5c5/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-joel-madden-good-charlotte-celeb.png",
+        "worth": "$2,246,400",
+        "worth_num": 2246400
+    },
+    {
+        "name": "Reese Witherspoon",
+        "wallet": "0x7948aa99e095dbfc1971bc8d2d1173893146630e",
+        "image": "https://dappradar.com/blog/static/25d40758c27506240da557cc0d83c463/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-reese-witherspoon-celeb-wallet.png",
+        "worth": "$25,570.53",
+        "worth_num": 25570.53
+    },
+    {
+        "name": "Jordan Belfort",
+        "wallet": "0xdbf2445e5049c04cda797dae60ac885e7d79df9d",
+        "image": "https://dappradar.com/blog/static/bec4cac32571184faab1ab3d9b12e7e9/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-jordan-belfort-celeb-wallet.png",
+        "worth": "$6,537.12",
+        "worth_num": 6537.12
+    },
+    {
+        "name": "DJ Marshmello",
+        "wallet": "0xa0eaf6b0df87132c9a28e450a43c1d906defb60b",
+        "image": "https://dappradar.com/blog/static/b9c2ce590bc8a8e32632c0b622b88ec0/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-dj-marshmello-celeb-wallet.png",
+        "worth": "$958,792.87",
+        "worth_num": 958792.87
+    },
+    {
+        "name": "Timbaland",
+        "wallet": "0x4b1bdae6b46c2ed904581d7e4bf2b71e5f3f7072",
+        "image": "https://dappradar.com/blog/static/6f68271d826b3536ef1b4bb7765710b5/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-timbaland-celeb-icon.png",
+        "worth": "$382.72",
+        "worth_num": 382.72
+    },
+    {
+        "name": "Lil Baby",
+        "wallet": "0xc86b12d850fdbbf3260a7baae862f85857aadbba",
+        "image": "https://dappradar.com/blog/static/c077d5d9433d1555d685a0585539fcea/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-lil-baby-button.png",
+        "worth": "$295,117.93",
+        "worth_num": 295117.93
+    },
+    {
+        "name": "Jimmy Fallon",
+        "wallet": "0x0394451c1238cec1e825229e692aa9e428c107d8",
+        "image": "https://dappradar.com/blog/static/75deed24d81cff7d63345fb60d63047e/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-jimmy-fallon.png",
+        "worth": "$319,079.04",
+        "worth_num": 319079.04
+    },
+    {
+        "name": "Post Malone",
+        "wallet": "0xbea020c3bd417f30de4d6bd05b0ed310ac586cc0",
+        "image": "https://dappradar.com/blog/static/4da2d5495e4d79334530c10d3b729e5a/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-post-malone.png",
+        "worth": "$631,170.62",
+        "worth_num": 631170.62
+    },
+    {
+        "name": "Logan Paul",
+        "wallet": "0xff0bd4aa3496739d5667adc10e2b843dfab5712b",
+        "image": "https://dappradar.com/blog/static/19f81b13bdbeb8d23aa07ed2470be25f/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-logan-paul.png",
+        "worth": "$850,185.4",
+        "worth_num": 850185.4
+    },
+    {
+        "name": "Gary Vee",
+        "wallet": "0xd6a984153acb6c9e2d788f08c2465a1358bb89a7",
+        "image": "https://dappradar.com/blog/static/0cee594e4856960e4b55c315665eab0f/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-gary-vaynerchuk-1.png",
+        "worth": "$2.07 Million",
+        "worth_num": 2070000
+    },
+    {
+        "name": "Paris Hilton",
+        "wallet": "0xb6aa5a1aa37a4195725cdf1576dc741d359b56bd",
+        "image": "https://dappradar.com/blog/static/3bcb9d6896384ef9b0fef02583a6d819/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-paris-hilton-gfinal.png",
+        "worth": "$558,775.68",
+        "worth_num": 558775.68
+    },
+    {
+        "name": "Neymar",
+        "wallet": "0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d",
+        "image": "https://dappradar.com/blog/static/559e2cd68cdfcc8bae7daf381683a3dd/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-neymar-1.png",
+        "worth": "Limitless",
+        "worth_num": null
+    },
+    {
+        "name": "Kevin Hart",
+        "wallet": "0xbbdac7ba85af15420afd1f4aa3313c3535b15cde",
+        "image": "https://dappradar.com/blog/static/c9a040e0cba8954bb87621af765c38ef/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-kevin-hart.png",
+        "worth": "$216,558.12",
+        "worth_num": 216558.12
+    },
+    {
+        "name": "Gwyneth Paltrow",
+        "wallet": "0x31185f782a7c11044566d70dfcf1c8175486f451",
+        "image": "https://dappradar.com/blog/static/c9c3c5436097589b9611c378efae9785/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-gwyn.png",
+        "worth": "$224,307.46",
+        "worth_num": 224307.46
+    },
+    {
+        "name": "Eva Longoria",
+        "wallet": "0xaa1b056286a66a9e6752c26776ac034c662a51d5",
+        "image": "https://dappradar.com/blog/static/9a84e87988d0402f1a2005c3c85f54f0/e08cd/dappradar.com-celebrity-wallets-a-dive-into-crypto-hollywood-eva-longoria-butotn.png",
+        "worth": "$32,071.52",
+        "worth_num": 32071.52
+    },
+    {
+        "name": "Dillon Francis",
+        "wallet": "0x58473e9ac681c4424ca74619281ff71801d002d6",
+        "image": "https://dappradar.com/blog/static/1ca1806ced7b11fe7a8868909e707a73/95fae/dappradar.com-celebrity-nft-wallets-dillon-francis-block.png",
+        "worth": "$478,494.85",
+        "worth_num": 478494.85
+    },
+    {
+        "name": "Madonna",
+        "wallet": "0x58473e9ac681c4424ca74619281ff71801d002d6",
+        "image": "https://dappradar.com/blog/static/6ac0b4c888d7f45cef3e42f8e7bd3366/95fae/dappradar.com-celebrity-nft-wallets-madonna-thumb.png",
+        "worth": "$1,763.77",
+        "worth_num": 1763.77
+    }
+  ];
+
+  export default whaleTable
