+++ client/store.js
@@ -1,11 +1,14 @@
-import { createStore } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
+import thunkMiddleware from 'redux-thunk'
 import reducers from './reducers/index';
+// import thunk and applyMiddleware here?
 
 // we are adding composeWithDevTools here to get easy access to the Redux dev tools
 const store = createStore(
   reducers,
-  composeWithDevTools()
+  composeWithDevTools(applyMiddleware(thunkMiddleware))
+  // createMiddleware()
 );
 
 export default store;
