+++ package.json
@@ -62,6 +62,7 @@
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
+    "redux-thunk": "^2.4.1",
     "webpack": "^5.73.0"
   }
 }
